refactor AxisLimits2D and AxisDimensions
